+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/vungle/mediation/VungleInterstitialAdapter.java
@@ -2,7 +2,6 @@ package com.vungle.mediation;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.util.Log;
 
 import com.google.android.gms.ads.AdRequest;
 import com.google.android.gms.ads.mediation.MediationAdRequest;
@@ -14,9 +13,6 @@ import com.vungle.publisher.AdConfig;
  * A {@link MediationInterstitialAdapter} used to load and show Vungle interstitial ads using
  * Google Mobile Ads SDK mediation.
  */
-import java.util.ArrayList;
-import java.util.List;
-
 public class VungleInterstitialAdapter implements MediationInterstitialAdapter {
 
     private static final String TAG = VungleManager.class.getSimpleName();
@@ -56,8 +52,7 @@ public class VungleInterstitialAdapter implements MediationInterstitialAdapter {
         void onInitialized(boolean isSuccess) {
             if (mMediationInterstitialListener != null) {
                 if (!isSuccess) {
-                    mMediationInterstitialListener.onAdFailedToLoad(
-                            VungleInterstitialAdapter.this,
+                    mMediationInterstitialListener.onAdFailedToLoad(VungleInterstitialAdapter.this,
                             AdRequest.ERROR_CODE_INTERNAL_ERROR);
                 } else {
                     loadAd();
@@ -82,9 +77,11 @@ public class VungleInterstitialAdapter implements MediationInterstitialAdapter {
                                       MediationAdRequest mediationAdRequest,
                                       Bundle mediationExtras) {
         try {
-            AdapterParametersParser.Config config = AdapterParametersParser.parse(mediationExtras, serverParameters);
+            AdapterParametersParser.Config config =
+                    AdapterParametersParser.parse(mediationExtras, serverParameters);
             mMediationInterstitialListener = mediationInterstitialListener;
-            mVungleManager = VungleManager.getInstance(config.getAppId(), config.getAllPlacements());
+            mVungleManager =
+                    VungleManager.getInstance(config.getAppId(), config.getAllPlacements());
 
             mPlacementForPlay = mVungleManager.findPlacement(mediationExtras, serverParameters);
             mAdConfig = VungleExtrasBuilder.adConfigWithNetworkExtras(mediationExtras);
@@ -99,9 +96,12 @@ public class VungleInterstitialAdapter implements MediationInterstitialAdapter {
                 mVungleListener.setWaitingInit(true);
                 mVungleManager.init(context);
             }
-        } catch (IllegalArgumentException _) {
-            if (mediationInterstitialListener != null)
-                mediationInterstitialListener.onAdFailedToLoad(VungleInterstitialAdapter.this, AdRequest.ERROR_CODE_INVALID_REQUEST);
+        } catch (IllegalArgumentException e) {
+            if (mediationInterstitialListener != null) {
+                mediationInterstitialListener
+                        .onAdFailedToLoad(VungleInterstitialAdapter.this,
+                                AdRequest.ERROR_CODE_INVALID_REQUEST);
+            }
         }
     }