+++ ThirdPartyAdapters/applovin/CHANGELOG.md
@@ -1,5 +1,17 @@
 # AppLovin Adapter for Google Mobile Ads SDK for Android
 
+## 7.6.2.0
+- Verified compatibility with AppLovin SDK 7.6.2.
+
+## 7.6.1.0
+- Verified compatibility with AppLovin SDK 7.6.1.
+
+## 7.6.0.0
+- Verified compatibility with AppLovin SDK 7.6.0.
+
+## 7.5.0.0
+- Verified compatibility with AppLovin SDK 7.5.0.
+
 ## 7.4.1.1
 - Added support for banner ads.