+++ ThirdPartyAdapters/mopub/CHANGELOG.md
@@ -1,5 +1,8 @@
 # MoPub Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 4.19.0.0
+- Verified compatibility with MoPub SDK 4.19.0.
+
 ## 4.18.0.0
 - Verified compatibility with MoPub SDK 4.18.0.
 - Updated the Adapter project for Android Studio 3.0 and Android API 26.