+++ ThirdPartyAdapters/nend/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Nend Adapter for Google Mobile Ads SDK for Android Changelog
 
-## x.y.z.0
+## 4.0.2.0
 
-- First release!
-- Adds support for banners, interstitials, and rewarded, etc.
+- First release in Google Mobile Ads Mediation open source project.
+- Added support for banner, interstitial, and rewarded video ads.